Add files via upload
I completely overhauled the model.  I have worked diligently on it, but still can't get it to go.  There are presently no TA's available, and they won't have office hours until after the Exam codes are released, when they are unlikely to respond due to exam blackout protocols.

I hope that the dozens of hours I've spent working on this will get me close enough to be able to unlock my exam.